requirements/base.txt
Django>=5.0,<6.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-environ>=0.11
django-filter>=24.1
drf-spectacular>=0.27
django-cors-headers>=4.3
psycopg[binary]>=3.1
celery>=5.3
django-celery-beat>=2.6
django-celery-results>=2.5
structlog>=24.1
django-guardian>=2.4
django-fsm>=2.8
